requirements.txt
google-generativeai
streamlit
scikit-learn
